Break constants into lifeCycle and permissions
