+++ libs/wcm/constants/lifeCycle.js
@@ -1,0 +1,19 @@
+export const PAIRING_PROPOSAL_STATUS = {
+  ERROR: 'ERROR',
+  SUCCESS: 'SUCCESS',
+};
+
+export const EVENTS = {
+  SESSION_PROPOSAL: 'session_proposal',
+  SESSION_REQUEST: 'session_request',
+  SESSION_PING: 'session_ping',
+  SESSION_EVENT: 'session_event',
+  SESSION_UPDATE: 'session_update',
+  SESSION_DELETE: 'session_delete',
+};
+
+export const ERROR_CASES = {
+  USER_DISCONNECTED: 'USER_DISCONNECTED',
+  INVALID_METHOD: 'INVALID_METHOD',
+  USER_REJECTED_METHODS: 'USER_REJECTED_METHODS',
+};
