+++ libs/wcm/constants/permissions.js
@@ -1,0 +1,9 @@
+export const LISK_SIGNING_METHODS = {
+  LISK_SIGN_TRANSACTION: 'lisk_signTransaction',
+  LISK_SIGN_MESSAGE: 'lisk_signMessage',
+};
+
+export const LISK_EVENTS = {
+  LISK_TRANSACTION_MADE: 'lisk_transaction_made',
+  LISK_TOKEN_RECEIVED: 'lisk_token_received',
+};
